+++ server/src/controllers/global_data_sources.controller.ts
@@ -13,19 +13,20 @@ import {
 } from '@nestjs/common';
 import { JwtAuthGuard } from '../../src/modules/auth/jwt-auth.guard';
 import { decamelizeKeys } from 'humps';
-import { AppsService } from '@services/apps.service';
 import { AppsAbilityFactory } from 'src/modules/casl/abilities/apps-ability.factory';
+import { GlobalDataSourceAbilityFactory } from 'src/modules/casl/abilities/global-datasource-ability.factory';
 import { DataQueriesService } from '@services/data_queries.service';
 import { GlobalDataSourcesService } from '@services/global_data_sources.service';
 import { AuthorizeDataSourceOauthDto, CreateDataSourceDto, UpdateDataSourceDto } from '@dto/data-source.dto';
 import { decode } from 'js-base64';
 import { User } from 'src/decorators/user.decorator';
+import { DataSource } from 'src/entities/data_source.entity';
 
 @Controller('v2/data_sources')
 export class GlobalDataSourcesController {
   constructor(
-    private appsService: AppsService,
     private appsAbilityFactory: AppsAbilityFactory,
+    private globalDataSourceAbilityFactory: GlobalDataSourceAbilityFactory,
     private globalDataSourcesService: GlobalDataSourcesService,
     private dataQueriesService: DataQueriesService
   ) {}
@@ -51,31 +52,19 @@ export class GlobalDataSourcesController {
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  async create(@User() user, @Query('environment_id') environmentId, @Body() createDataSourceDto: CreateDataSourceDto) {
-    const { kind, name, options, plugin_id: pluginId, scope } = createDataSourceDto;
-
-    const dataSource = await this.globalDataSourcesService.create(
-      name,
-      kind,
-      options,
-      user.organizationId,
-      scope,
-      pluginId,
-      environmentId
-    );
-    return decamelizeKeys(dataSource);
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post()
   async createGlobalDataSources(@User() user, @Body() createDataSourceDto: CreateDataSourceDto) {
     const { kind, name, options, plugin_id: pluginId, scope } = createDataSourceDto;
 
+    const ability = await this.globalDataSourceAbilityFactory.globalDataSourceActions(user);
+
+    if (!ability.can('createGlobalDataSource', DataSource)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
+
     const dataSource = await this.globalDataSourcesService.create(
       name,
       kind,
       options,
-      null,
       user.organizationId,
       scope,
       pluginId
@@ -92,7 +81,12 @@ export class GlobalDataSourcesController {
     @Query('environment_id') environmentId,
     @Body() updateDataSourceDto: UpdateDataSourceDto
   ) {
-    console.log({ idforog: user.organizationId, user });
+    const ability = await this.globalDataSourceAbilityFactory.globalDataSourceActions(user);
+
+    if (!ability.can('updateGlobalDataSource', DataSource)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
+
     const { name, options } = updateDataSourceDto;
 
     await this.globalDataSourcesService.update(dataSourceId, user.organizationId, name, options, environmentId);
@@ -102,6 +96,11 @@ export class GlobalDataSourcesController {
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
   async delete(@User() user, @Param('id') dataSourceId) {
+    const ability = await this.globalDataSourceAbilityFactory.globalDataSourceActions(user);
+
+    if (!ability.can('deleteGlobalDataSource', DataSource)) {
+      throw new ForbiddenException('You do not have permissions to perform this action');
+    }
     const result = await this.globalDataSourcesService.delete(dataSourceId);
     if (result.affected == 1) {
       return;
@@ -127,7 +126,7 @@ export class GlobalDataSourcesController {
     const ability = await this.appsAbilityFactory.appsActions(user, app.id);
 
     if (!ability.can('authorizeOauthForSource', app)) {
-      throw new ForbiddenException('you do not have permissions to perform this action');
+      throw new ForbiddenException('You do not have permissions to perform this action');
     }
 
     await this.dataQueriesService.authorizeOauth2(dataSource, code, user.id, environmentId);
